Use long double as the type of possibiliy.

+++ include/kernels/outside.cuh
@@ -15,8 +15,9 @@
 #ifdef USE_CUDA
 __global__
 #endif
-void kernel_outside_main(float* mu, float* beta, const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
-                        uint32_t* grammar_index, uint32_t* grammar_table, float* alpha, 
+void kernel_outside_main(long double* mu, long double* beta, 
+                        const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
+                        uint32_t* grammar_index, uint32_t* grammar_table, long double* alpha, 
                         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars,
                         std::vector<std::tuple<uint32_t, uint32_t>> inside_order_1_rule_iteration_path
                         #ifdef DEBUG_INSIDE_ALGORITHM
